Fix: Cast SharedCacheSize and LocalCacheSize to ulong in NoWxCache constructor
Changed from const int to static readonly int broke implicit conversion to ulong.
Explicit casts are now required when passing to MemoryCache constructor.

+++ src/Ryujinx.Cpu/LightningJit/Cache/NoWxCache.cs
@@ -151,8 +151,8 @@ namespace Ryujinx.Cpu.LightningJit.Cache
         {
             _stackWalker = stackWalker;
             _translator = translator;
-            _sharedCache = new(allocator, SharedCacheSize);
-            _localCache = new(allocator, LocalCacheSize);
+            _sharedCache = new(allocator, (ulong)SharedCacheSize);
+            _localCache = new(allocator, (ulong)LocalCacheSize);
             _pendingMap = new(_sharedCache.ReprotectAsRx, RegisterFunction);
         }
